requirements.txt
streamlit
langchain
langchain-openai
langchain-community
openai
pypdf
python-dotenv
matplotlib
pandas
Pillow